fixed Iservice interface for integration with UI

+++ Project0/RestaurantService/IService.cs
@@ -2,7 +2,7 @@ namespace RestaurantService
 {
     public interface IService
     {
-        void CalculateAvg(List<decimal> ratings);
+        void CalculateAvg(List<double> ratings);
 
 
     }
